move KeypadMnemonics file into keyboard package
