+++ src/org/tn5250j/keyboard/KeypadMnemonic.java
@@ -22,7 +22,7 @@
  * the Free Software Foundation, Inc., 59 Temple Place, Suite 330,
  * Boston, MA 02111-1307 USA
  */
-package org.tn5250j.sessionsettings;
+package org.tn5250j.keyboard;
 
 public enum KeypadMnemonic {
