+++ src/org/tn5250j/keyboard/actions/AttributesAction.java
@@ -32,7 +32,7 @@ import javax.swing.*;
 import java.awt.event.ActionEvent;
 import java.awt.event.KeyEvent;
 
-import static org.tn5250j.sessionsettings.KeypadMnemonic.DISP_ATTRIBUTES;
+import static org.tn5250j.keyboard.KeypadMnemonic.DISP_ATTRIBUTES;
 
 /**
  * Display session attributes
