+++ src/org/tn5250j/keyboard/actions/CopyAction.java
@@ -32,7 +32,7 @@ import javax.swing.*;
 import java.awt.event.ActionEvent;
 import java.awt.event.KeyEvent;
 
-import static org.tn5250j.sessionsettings.KeypadMnemonic.COPY;
+import static org.tn5250j.keyboard.KeypadMnemonic.COPY;
 
 /**
  * Copy action
