+++ src/org/tn5250j/keyboard/actions/GuiAction.java
@@ -33,7 +33,7 @@ import javax.swing.*;
 import java.awt.event.ActionEvent;
 import java.awt.event.KeyEvent;
 
-import static org.tn5250j.sessionsettings.KeypadMnemonic.GUI;
+import static org.tn5250j.keyboard.KeypadMnemonic.GUI;
 
 /**
  * Toggle gui
