+++ src/org/tn5250j/keyboard/actions/PrintAction.java
@@ -32,7 +32,7 @@ import javax.swing.*;
 import java.awt.event.ActionEvent;
 import java.awt.event.KeyEvent;
 
-import static org.tn5250j.sessionsettings.KeypadMnemonic.PRINT_SCREEN;
+import static org.tn5250j.keyboard.KeypadMnemonic.PRINT_SCREEN;
 
 /**
  * Display session attributes
