+++ src/org/tn5250j/keyboard/actions/RulerAction.java
@@ -32,7 +32,7 @@ import javax.swing.*;
 import java.awt.event.ActionEvent;
 import java.awt.event.KeyEvent;
 
-import static org.tn5250j.sessionsettings.KeypadMnemonic.CURSOR;
+import static org.tn5250j.keyboard.KeypadMnemonic.CURSOR;
 
 /**
  * Display session attributes
